Rename easse/utils/paths.py to easse/utils/constants.py
