+++ easse/utils/constants.py
